requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
praw>=7.7.1
nltk>=3.8.1
textblob>=0.17.1
pandas>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
